Flatten initial development
